:recycle: (FE:WebSocket) セッションと連携するため、使い方の変更
